requirements.txt
Django==1.11.6
django-bootstrap3==9.0.0
django-debug-toolbar==1.8
django-filter==1.1.0
django-pdb==0.6.1
pytz==2017.2
sqlparse==0.2.4
virtualenv==15.1.0
